feat: 印刷時にdetailsを開く処理をコンポーネントとして実装
